remove and rename file
